extract nxtp handling
