+++ .env
@@ -3,11 +3,11 @@ REACT_APP_PROJECT_NAME=Li.Finance
 REACT_APP_ANALYTICS_ID=
 REACT_APP_API_URL=
 
-REACT_APP_RPC_URL_MAINNET="https://mainnet.infura.io/v3/b080ce70cc7e44fdbfcc6874149f5a63"
-REACT_APP_RPC_URL_ROPSTEN="https://ropsten.infura.io/v3/b080ce70cc7e44fdbfcc6874149f5a63"
-REACT_APP_RPC_URL_RINKEBY="https://rinkeby.infura.io/v3/b080ce70cc7e44fdbfcc6874149f5a63"
-REACT_APP_RPC_URL_KOVAN="https://kovan.infura.io/v3/b080ce70cc7e44fdbfcc6874149f5a63"
-REACT_APP_RPC_URL_GORLI="https://goerli.infura.io/v3/b080ce70cc7e44fdbfcc6874149f5a63"
+REACT_APP_RPC_URL_MAINNET="https://mainnet.infura.io/v3/9dfee2b92a8041f7b1254445d5b58763"
+REACT_APP_RPC_URL_ROPSTEN="https://ropsten.infura.io/v3/9dfee2b92a8041f7b1254445d5b58763"
+REACT_APP_RPC_URL_RINKEBY="https://rinkeby.infura.io/v3/9dfee2b92a8041f7b1254445d5b58763"
+REACT_APP_RPC_URL_KOVAN="https://kovan.infura.io/v3/9dfee2b92a8041f7b1254445d5b58763"
+REACT_APP_RPC_URL_GORLI="https://goerli.infura.io/v3/9dfee2b92a8041f7b1254445d5b58763"
 
 REACT_APP_RPC_URL_BSC="https://bsc-dataseed1.defibit.io"
 # or "https://bsc-dataseed.binance.org/"
