+++ .env.nxtp
@@ -10,3 +10,10 @@ REACT_APP_METADATA_DESCRIPTION="Transfer token between testnets"
 REACT_APP_METADATA_IMAGE_URL="https://nxtp.li.finance/powered_by_connext_lifinance.png"
 REACT_APP_METADATA_IMAGE_WIDTH="589"
 REACT_APP_METADATA_IMAGE_HEIGHT="167"
+
+# RPCs
+REACT_APP_RPC_URL_MAINNET="https://mainnet.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
+REACT_APP_RPC_URL_ROPSTEN="https://ropsten.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
+REACT_APP_RPC_URL_RINKEBY="https://rinkeby.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
+REACT_APP_RPC_URL_KOVAN="https://kovan.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
+REACT_APP_RPC_URL_GORLI="https://goerli.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
