+++ .env.production
@@ -1,3 +1,10 @@
 REACT_APP_ANALYTICS_ID=G-7C14ZNL54Q
 REACT_APP_API_URL="https://li.finance/api/"
 REACT_APP_STOMT_ENABLED=true
+
+# RPCs
+REACT_APP_RPC_URL_MAINNET="https://mainnet.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
+REACT_APP_RPC_URL_ROPSTEN="https://ropsten.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
+REACT_APP_RPC_URL_RINKEBY="https://rinkeby.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
+REACT_APP_RPC_URL_KOVAN="https://kovan.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
+REACT_APP_RPC_URL_GORLI="https://goerli.infura.io/v3/c7fe4abb5bbc466cb56038efbb0b9930"
